requirements.txt
aiosqlite==0.20.0
aiohttp==3.10.5
python-telegram-bot==21.5
pytgcalls==3.0.0
redis==5.0.8
telethon==1.36.0
websockets==13.0.1
yt-dlp==2024.8.6
